scheduler-service/requirements.txt
asgiref==3.6.0
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==3.2
django-3-jet==1.0.8
falcon==3.1.1
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
mysqlclient==2.1.1
python-dateutil~=2.8.2
apscheduler==3.10.0
beautifulsoup4~=4.11.2